refactor(auth): split index.ts into plugin.ts and acl module
- Create src/auth/acl.ts with allowAccess, allowPublish, allowUnpublish
- Create src/plugin.ts with AuthCustomPlugin class (default export)
- Hollow src/index.ts to single re-export: export { default } from './plugin'
- Empty-password guards remain in plugin.ts authenticate() as synchronous early returns
- All 132 pre-existing tests pass; coverage at or above baseline
